backend/requirements.txt
fastapi
uvicorn
supabase
qdrant-client
openai
pypdf
python-docx
python-dotenv
requests
pydantic-settings
sentence-transformers